removes XMLParser subclass
